sep 8

+++ House.cs
@@ -1,9 +1,6 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
+﻿
 public class House : Building {
-	public int housing = 2;
+	public int housing { get; protected set; }
 
 	public override void SetBasement(SurfaceBlock b, PixelPosByte pos) {		
 		if (b == null) return;
@@ -12,6 +9,18 @@ public class House : Building {
 	protected void PrepareHouse(SurfaceBlock b, PixelPosByte pos) {
 		SetBuildingData(b,pos);
 		GameMaster.colonyController.AddHousing(this);
+        switch (id)
+        {
+            case LANDED_ZEPPELIN_ID: housing = 10;break;
+            case HQ_2_ID: housing = 30;break;
+            case HQ_3_ID: housing = 40;break;
+            case HQ_4_ID: housing = 45;break;
+            case HOUSE_0_ID: housing = 2; break;
+            case HOUSE_1_ID: housing = 10;break;
+            case HOUSE_2_ID: housing = 50;break;
+            case HOUSE_3_ID: housing = 100;break;
+            case HOUSE_5_ID: housing = 800;break;
+        }
 	}
 
 	override public void SetActivationStatus(bool x) {
@@ -20,9 +29,11 @@ public class House : Building {
 		GameMaster.colonyController.RecalculateHousing();
 		ChangeRenderersView(x);
 	}
-
-	 void OnDestroy() {
-		PrepareBuildingForDestruction();
-		GameMaster.colonyController.DeleteHousing(this);
-	}
+    override public void Annihilate(bool forced)
+    {
+        if (destroyed) return;
+        else destroyed = true;
+        PrepareBuildingForDestruction(forced);
+        GameMaster.colonyController.DeleteHousing(this);
+    }
 }
